+++ __tests__/addon-registry.test.js
@@ -27,10 +27,20 @@ describe('AddonConfigurationRegistry', () => {
       'test-released-addon',
       'test-released-source-addon',
       'my-volto-config-addon',
+      'test-released-dummy',
       'test-released-unmentioned',
     ]);
 
     expect(reg.packages).toEqual({
+      'non-volto-addon-lib': {
+        addons: [],
+        isPublishedPackage: false,
+        isRegisteredAddon: false,
+        modulePath: `${base}/addons/non-volto-addon-lib/src`,
+        name: 'non-volto-addon-lib',
+        packageJson: `${base}/addons/non-volto-addon-lib/package.json`,
+        version: '0.0.0',
+      },
       'test-addon': {
         isPublishedPackage: false,
         modulePath: `${base}/addons/test-addon/src`,
@@ -83,7 +93,7 @@ describe('AddonConfigurationRegistry', () => {
         modulePath: `${base}/addons/test-released-dummy`,
         name: 'test-released-dummy',
         packageJson: `${base}/addons/test-released-dummy/package.json`,
-        isRegisteredAddon: false,
+        isRegisteredAddon: true,
         version: '0.0.0',
       },
     });
@@ -93,6 +103,7 @@ describe('AddonConfigurationRegistry', () => {
     const base = path.join(__dirname, 'fixtures', 'test-volto-project');
     const reg = new AddonConfigurationRegistry(base);
     expect(reg.getResolveAliases()).toStrictEqual({
+      'non-volto-addon-lib': `${base}/addons/non-volto-addon-lib/src`,
       'my-volto-config-addon': `${base}/addons/my-volto-config-addon/src`,
       'test-addon': `${base}/addons/test-addon/src`,
       'test-released-addon': `${base}/node_modules/test-released-addon`,